# TINKERPOP-915 Add combined handler for Http and Websockets