@spmallette 2017-06-06
This is really great stuff. I tried doing this a few times and for whatever reason I couldn't get it to work right. Looking forward to reviewing this one.