@spmallette 2017-06-15
Sorry @krlohnes - if you're not following along we had a snag in 3.2.5 release so the release branches still aren't open. Probably won't see this PR handled until next week some time at this rate.