@krlohnes 2017-06-19
@spmallette Yeah, I'm on the mailing list, I've been following along. No worries.