@spmallette 2017-06-19
`GremlinServerWsAndHttpIntegrateTest.java` is a nice validation that all the REST tests work and a reasonable portion websockets tests work, but is there a nicer way to do that? Under this model, every time we add a new test we have to also implement that test method in `GremlinServerWsAndHttpIntegrateTest.java`. We're also not really testing all functionality of websockets as it only does `GremlinServerIntegrateTest`. 

anyway - any thoughts on a better testing model?