@krlohnes 2017-06-19
I can probably figure something out with JUnit `Parameters` for the `GremlinServerIntegrateTest` and `GremlinServerHttpIntegrateTest`. I'll have a go at it.