@spmallette 2017-06-19
Not sure how I feel about it but just throwing stuff out there to think about....Another option might be to just have a smoke test for the combined channelizer? or maybe all tests should run over your combined channelizer and we just need smoke tests for the specific ones? do you feel like smoke tests are too risky given your changes? Do we need to re-test all functionality over all the channelizers?