@krlohnes 2017-06-19
My change can be summed up by 

"Setup the WebSocket stuff and insert the `GremlinEndpointHandler` into the `ChannelPipeline` to serve Http requests when necessary"

Changes to the `WebSocketChannelizer` and `SaslAuthenticationHandler` would likely be picked up easily enough by simply using the combined channelizer. If there was an issue with the `HttpChannelizer`, it may not be picked up by those tests.

Running all of the tests from both was mostly for my own sanity. I wanted to make sure I wasn't missing something.

If you're thinking about minimizing impact from extra tests running in the TP suite, I'd probably suggest running the combined channelizer on WS tests and both on `GremlinServerHttpIntegrateTest`.