@spmallette 2017-06-19
ok - let me think about it a bit before you write too much code in any particular direction. we'll also give others some time to chime in in case they have an opinion.