@spmallette 2017-06-22
The more i think about this the more I wonder if testing lots of server configurations with the same scenarios is useful.  Generally speaking the `Channelizer` isn't responsible for many of the features under test in `GremlinServerIntegrateTest` and there are other tests that are probably testing `Channelizer` capabilities that aren't represented there (maybe more true on master than tp32 - you might need a second PR when we sort this one out).

I think it would be worth identifying the tests that validate that a `Channelizer` is doing its job right and make a common set of tests that are executed over different server configurations. This would probably mean cherry picking tests from existing ones and moving them to this new test class.  

For example, `shouldBatchResultsByTwos` has nothing to do with the `Channelizer` selection really. That's an integration test for the `OpProcessor` really. On a similar note, `shouldReceiveFailureTimeOutOnScriptEval` is a integration test of the `GremlinExecutor` - such a test would not need to be executed across multiple configurations.

On the other hand, `shouldStartWithDefaultSettings` is quite generic and would be a good general smoke test. That one could be cherry picked to the new common test class and renamed to "shouldReturnListOfNumbers".  We'd probably also want to know if SSL/authentication can be enabled across all `Channelizer` implementations - that's a common feature and one controlled by the `Channelizer`.

Does that make sense to you at all?