@krlohnes 2017-06-23
Yeah, I definitely think that makes sense, and it's likely the simplest solution to the tests here. 

I'll make sure the new test hits `shouldStartWithDefaultSettings`, ssl and auth, and keep the test that toggles between the two protocols to make sure they can both be connected to. I think those would be a reasonable check on whether or not this functionality works. I'll ping you once I pick out those tests. If anything else comes to mind, let me know, and I'm happy to add it.