@spmallette 2017-06-27
@krlohnes have you done any performance testing to determine if the combined channelizer shows any difference to the independent channelizers? i think that would be worthwhile to know in evaluating this PR. As an added note which I didn't mention before I don't think - this PR will need reference docs, upgrade docs and changelog entries.