@krlohnes 2017-06-27
@spmallette I did some testing between the http channelizer and the combined one.

With Basic Auth, the mean latency with 10000 requests for the combined channelizer was 148.3 ms and the standalone channelizer was 148.9 ms.

Without Basic Auth the mean latency with 10000 request for the combined channelizer was 4.8ms and the standalone channelizer was 4.9 ms

Given the results for http and that the ChannelPipeline is only manipulated for http, I didn't run any performance tests for WebSockets.

I'll write up the docs once I'm done refactoring the Channelizer tests.