@spmallette 2017-06-27
interesting - thanks for that information.