@krlohnes 2017-06-29
@spmallete I refactored the channelizer tests and removed some elsewhere when I thought they were appropriately replaced.  I'm ultimately not all too familiar with Java NIO. I added it as another channelizer test, but I can delete that test class and the associated code in the abstract test class as it really only runs a single test. I wasn't quite sure what the best way to go with that was. I also added in the doc changes.