@spmallette 2017-06-30
I think NIO should have the same capabilities as the WS channelizer. I think you could setup NIO to test the same things WS does. 

Hate to nitpick but could you do a quick pass through your changes and `final` your variables as it's our style.

I like the "channel" tests. We really needed something unified like that.