@krlohnes 2017-06-30
@spmallette I fixed the review comments and I think I hit all the places I missed the `final` variables.