@krlohnes 2017-06-30
Okay. I got the NIOChannelizer tests passing now as well, things are `final`ed where appropriate, and I've addressed your PR comments.