@spmallette 2017-07-05
All tests pass with `docker/build.sh -t -n -i`

VOTE +1