@spmallette 2017-07-06
I fixed the broken test on tp32.  I guess this needs to be rebased again. Not sure how that failing test snuck through review....kinda odd