@spmallette 2017-07-10
@krlohnes looks like you have all the votes together for this now. it went into tp32 quite well of course, but when i tried to go to master, i hit some problems (tests failing). 

do you mind submitting a second separate pull request to master with these changes? if you do that i can get them both merged at the same time (no need more voting imo).