@krlohnes 2017-07-11
@spmallette From the other PR on master I'd had open

>Also, your PR looks a little odd - the commit id is not the same as the one on your other PR.

I'm not quite sure how to go about not mucking up the history here. I'd cherry-picked and fixed the merge conflicts/test failures but that changes the commit sha. 

I tried merging this branch with a new branch. That fixes the the commit id issue, but adds a merge commit. 

I'm a _little_ stumped on how to approach it from here.