@spmallette 2017-07-11
the merge commit is acceptable in this case. i'd even expect a follow on commit to fix test problems and such. `git merge` should give us as clean a history as we can get i this situation. thanks for asking on how to proceed.